Mod: Websocket processes moved to other file
