+++ src/main.ts
@@ -8,8 +8,8 @@
 import { serveDir, serveFile } from "jsr:@std/http@1";
 
 import { is_only_num_alp } from "./static/util/is_only_num_alp.js";
-import { encode_boolean_packet, PACKET_ID_MODERATOR_STATUS } from "./static/packet/boolean.js";
-import { encode_uint16_packet, PACKET_ID_PARTICIPANT_COUNT } from "./static/packet/uint16.js";
+import { ws_participant } from "./modules/ws_participant.ts";
+import { ws_moderator } from "./modules/ws_moderator.ts";
 
 
 /* Global variables */
@@ -50,39 +50,7 @@ Deno.serve(request => {
 
     // Main process of participant
     const { socket, response } = Deno.upgradeWebSocket(request);
-    const moderator_status_notice = () => socket.send(encode_boolean_packet(PACKET_ID_MODERATOR_STATUS, room_ids_moderator_connecting.has(room_id)));
-
-    //// Initial communication
-    socket.addEventListener("open", moderator_status_notice);
-
-    //// 'participant-count-update' (joined) signal to moderator
-    socket.addEventListener("open", () => {
-      const current_count = participant_count.get(room_id);
-      if(current_count)
-        participant_count.set(room_id, current_count + 1);
-      else
-        participant_count.set(room_id, 1);
-      event_core.dispatchEvent(new Event(`participant-count-update-${room_id}`));
-    });
-
-    //// 'participant-count-update' (left) signal to moderator
-    const participant_decrease_signal = () => {
-      const current_count = participant_count.get(room_id);
-      if(!current_count)
-        return; // Invalid case - participant left but count already 0 (undefined)
-      else if(current_count === 1)
-        participant_count.delete(room_id);
-      else
-        participant_count.set(room_id, current_count - 1);
-      event_core.dispatchEvent(new Event(`participant-count-update-${room_id}`));
-    }
-    socket.addEventListener("close", participant_decrease_signal);
-    socket.addEventListener("error", participant_decrease_signal);
-
-    //// 'moderator-update' signal from moderator
-    event_core.addEventListener(`moderator-update-${room_id}`, moderator_status_notice);
-    socket.addEventListener("close", () => event_core.removeEventListener(`moderator-update-${room_id}`, moderator_status_notice));
-    socket.addEventListener("error", () => event_core.removeEventListener(`moderator-update-${room_id}`, moderator_status_notice));
+    ws_participant(socket, room_id, event_core, room_ids_moderator_connecting, participant_count);
 
     return response;
 
@@ -100,30 +68,7 @@ Deno.serve(request => {
 
     // Main process of moderator
     const { socket, response } = Deno.upgradeWebSocket(request);
-    const participant_count_notice = () => socket.send(encode_uint16_packet(PACKET_ID_PARTICIPANT_COUNT, participant_count.get(room_id) || 0));
-    //                   When 'participant_count.get(room_id) === undefined' (means unset), set value 0  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-    //// Initial communication
-    socket.addEventListener("open", participant_count_notice);
-
-    //// 'participant-count-update' signal from participant
-    event_core.addEventListener(`participant-count-update-${room_id}`, participant_count_notice);
-    socket.addEventListener("close", () => event_core.removeEventListener(`participant-count-update-${room_id}`, participant_count_notice));
-    socket.addEventListener("error", () => event_core.removeEventListener(`participant-count-update-${room_id}`, participant_count_notice));
-
-    //// 'moderator-update' (joined) signal to participant
-    socket.addEventListener("open", () => {
-      room_ids_moderator_connecting.add(room_id);
-      event_core.dispatchEvent(new Event(`moderator-update-${room_id}`));
-    });
-
-    //// 'moderator-update' (left) signal to participant
-    const moderator_left_signal = () => {
-      room_ids_moderator_connecting.delete(room_id);
-      event_core.dispatchEvent(new Event(`moderator-update-${room_id}`));
-    }
-    socket.addEventListener("close", moderator_left_signal);
-    socket.addEventListener("error", moderator_left_signal);
+    ws_moderator(socket, room_id, event_core, room_ids_moderator_connecting, participant_count);
 
     return response;
